ARM: dts: msm: Change sde dsc linewidth to 2k resolution
Currently, the DSC linewidth is set to 2048 even though the hw can
support a width of 2560. The DP driver uses this number during mode
validation to determine the number of DSCs needed for that mode. So for
2k resolution, the DP Driver incorrectly reserves 2 DSCs but SDE driver
correctly enables only 1 DSC. This results in DP driver disabling DSC
on the second stream even though the resource is available.
This change updates the DSC linewidth to match hw support.

Change-Id: I38b2b3c5d28fd0f8c6229a6d46dba887372530b2