ARM: dts: msm: Introduce demura HFC region nodes
ABL expects the demura HFC region nodes to always be present
under a device's reserved memory regions. If this node is not
present, ABL won't be able to insert it and demura continuous
splash cannot be handled.
This change introduces dummy demura HFC regions into the device
tree. These nodes are given an address and size of 0 to allow
the kernel to specially handle them.

Change-Id: I69efe303e76070c6809fe7814a40b5bfe25d012b