ARM: dts: msm: Remove the redundant clocks from CSID node
CSID node has TFE and PHY src clocks. This causes the change in
clock rates of phy and TFE without being the driver aware of it.

This commit removes the other driver clocks from CSID. This will
ensure that only the respective driver is responsible for setting
any clock rate.

CRs-Fixed: 3264555
Change-Id: Ifd8ec7c4c4573761b7363543052c79fda214e97c