ARM: dts: qcom: Add wsa_data_fs_ctl_reg address in dtsi
Add wsa_data_fs_ctl_reg address in dtsi which will
be used when WSA2 backend is used.
By default this register is set to value "0", which
will enable the wsa_rx4 and wsa_rx5 and used to drive
slaves connected over wsa macro.
In order to enable wsa2_rx4 and wsa2_rx5 dev to drive
slaves connected over wsa2 macro we need to program
this register to "1".

Change-Id: Iec3af3780199e8ce9a2426104313291809d7bfbf