ARM: dts: qcom: Reorder ssr dev register seq for blair
It is expected that pinctrl-lpi gets the service up
notification before bolero ssr handler/notifier chain
gets triggered. This is needed as the pinctrl-lpi driver
is a dependency for the bolero and other soundwire codecs.

Updated sampling rate in the Amic configs to fix the noise
in Amic2 Tx path.

Change-Id: I44b5dcac1cd23691aef2a46ca5f6c83d118b9afb