ARM: dts: msm: update display ramdump memory size for volcano target
Continuous splash memory and display ramdump use the same DDR
address space. After the splash handoff commit, memory reserved
for the splash screen will be freed, except for the portion
allocated for the display ramdump. To achieve this, update the
display ramdump memory block size to 8MB as per the panel
resolution so the rest of the splash memory can be freed.

Change-Id: I73ee15c3ac2417bb67bcaa02d5b5df25200c4ba8